# Refactor SplitCamelCaseWord method