@skoshelev 2024-08-08
Hi @holydk! This method converts values ​​from enumerations if they do not have a locale in the DB. It would be interesting to see the difference in performance when parsing, say, the ReadonlyCheckboxes value.