@holydk 2024-08-08
Hi @skoshelev. I just saw [Added new ACL rules](https://github.com/nopSolutions/nopCommerce/commit/40692e04e9cf3dc6ad2148008576ef78626221fb) commit added that changed the method and decided to do a little optimization.