@skoshelev 2024-08-19
Hi @holydk. I understand, but I think that such optimization is unnecessary, the code readability will deteriorate and the real performance will not change. In addition, then, logically, it will be necessary to rewrite another 34 places where the Aggregate method is used. But it seems to me that in this case, readability is still a priority.